Feature/single threaded (#21)
* Add Linux release with debug info preset

* Remove locking primitives

These primitives were not used in the current software build and slowed down the program.

* Swap to logCompare script for nestest

logCompare.py provides better error messages on failure

* Remove WORKING_DIRECTORY
CMake implicitly uses CMAKE_CURRENT_BINARY_DIR, which works for our purposes

* Remove unnecessary heap allocation

* Avoid creating string on each loop

Instead, store the data for creating the string and write it out to the log file later on.

* Swap to std::filesystem for ROM path
This allows us to more easily normalize paths in the future, if desired.

* Fix bug w uninitialized nmiState
The nmiState flag was set to a garbage value on startup, leading to inconsistent behavior when running the executable.

* Adjust comment

* Fix log order

* Remove comment

* Address PR comments